split overlay from interaction view
